+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Improvements:
 - Added execute examples in help flag
 - Added `bp` alias in the `build-push` subcommand
 - Renamed the `function remove` command to` function undeploy`
+- Added flag namespace in the `function undeploy` subcommand
 
 ## Version v0.0.7 - 2020-07-03